TG-52: clean code
