+++ src/Components/Share/Form/form.module.css
@@ -1,4 +1,4 @@
-/* * {
+* {
   box-sizing: border-box;
 }
 
@@ -21,7 +21,7 @@
   display: flex;
   flex-wrap: wrap;
   justify-content: space-between;
-  width: 60vw;
+  width: 62vw;
   column-gap: 10px;
 }
 
@@ -76,11 +76,11 @@
   background: #FFFFFF;
   border: 1px solid #373867;
   border-radius: 5px;
-} */
+}
 
 /*Team Members*/
 
-/* .formSection form > .teamMembers {
+.formSection form > .teamMembers {
   width: 100%;
   max-height: 15vw;
   margin-top: 1vw;
@@ -141,4 +141,4 @@
 .teamMembers td img {
   width: 1.5vw;
   cursor: pointer;
-} */
+}
