+++ src/redux/admins/reducer.js
@@ -2,12 +2,6 @@ import {
   GET_ADMINS_PENDING,
   GET_ADMINS_SUCCESS,
   GET_ADMINS_ERROR,
-  // PUT_ADMINS_PENDING,
-  // PUT_ADMINS_SUCCESS,
-  // PUT_ADMINS_ERROR,
-  // POST_ADMINS_PENDING,
-  // POST_ADMINS_SUCCESS,
-  // POST_ADMINS_ERROR,
   DELETE_ADMINS_PENDING,
   DELETE_ADMINS_SUCCESS,
   DELETE_ADMINS_ERROR
