+++ src/routes/employees.jsx
@@ -1,15 +1,13 @@
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Employees from 'Components/Employees';
-// import Form from 'Components/Share/Form';
-import CreateForm from 'Components/Share/CreateForm';
+import Form from 'Components/Share/Form';
 
 const EmployeesRoutes = () => {
   return (
     <Router>
       <Switch>
         <Route exact path="/employees" component={Employees} />
-        <Route path="/employees/form/:id" render={() => <CreateForm />} />
-        <Route path="/employees/new" render={() => <CreateForm />} />
+        <Route path="/employees/form/:id" render={() => <Form />} />
       </Switch>
     </Router>
   );
