+++ src/routes/projects.jsx
@@ -1,15 +1,13 @@
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Projects from 'Components/Projects';
-// import Form from 'Components/Share/Form';
-import CreateForm from 'Components/Share/CreateForm';
+import Form from 'Components/Share/Form';
 
 const ProjectsRoutes = () => {
   return (
     <Router>
       <Switch>
         <Route exact path="/projects" component={Projects} />
-        <Route path="/projects/form/:id" render={() => <CreateForm />} />
-        <Route path="/projects/new" render={() => <CreateForm />} />
+        <Route path="/projects/form/:id" render={() => <Form />} />
       </Switch>
     </Router>
   );
