+++ src/routes/time-sheets.jsx
@@ -1,15 +1,13 @@
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import TimeSheets from 'Components/TimeSheets';
-// import Form from 'Components/Share/Form';
-import CreateForm from 'Components/Share/CreateForm';
+import Form from 'Components/Share/Form';
 
 const TimeSheetsRoutes = () => {
   return (
     <Router>
       <Switch>
         <Route exact path="/time-sheets" component={TimeSheets} />
-        <Route path="/time-sheets/form/:id" render={() => <CreateForm />} />
-        <Route path="/time-sheets/new" render={() => <CreateForm />} />
+        <Route path="/time-sheets/form/:id" render={() => <Form />} />
       </Switch>
     </Router>
   );
